backend/requirements.txt
fastapi
uvicorn
sqlalchemy
openai
faiss-cpu
transformers
torch
python-dotenv
markdown
numpy
tiktoken
protobuf
sentencepiece
python-multipart
psycopg2-binary